+++ app/components/users/paywalled_component.rb
@@ -1,15 +1,16 @@
 module Users
   class PaywalledComponent < ApplicationComponent
-    def initialize(user:, paywalled:, size: nil, title: nil, description: nil)
+    def initialize(user:, paywalled:, size: nil, title: nil, description: nil, public_key: nil)
       @user = user
       @paywalled = paywalled
       @size = size
       @title = title
       @description = description
+      @public_key = public_key
     end
 
     def render_content?
-      customer? || owner?
+      customer? || owner? || valid_public_profile_access?
     end
 
     def small?
@@ -36,6 +37,10 @@ module Users
 
     def owner?
       @paywalled&.user == @user && @user.present?
+    end
+
+    def valid_public_profile_access?
+      @paywalled&.valid_public_profile_access?(@paywalled, @public_key)
     end
   end
 end