+++ test/integration/developers_test.rb
@@ -363,6 +363,15 @@ class DevelopersTest < ActionDispatch::IntegrationTest
     assert_text developers(:one).hero
   end
 
+  test "developer hidden profile information is rendered with public profile key" do
+    sign_in users(:empty)
+    developer = developers(:one)
+    developer.share_url
+    get developer_public_url(developer, developer.public_profile_key)
+
+    refute_text I18n.t("subscription_cta_component.title")
+  end
+
   def assert_text(text)
     assert_select "*", text:
   end