+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 ### September
 
+* September 22  - Display developer private information with publicly shared URL #646 @sarvaiyanidhi
 * September 21 - Only show page 1 of search results to non-subscribers #661 @mansakondo
 * September 18 - Add badges to developer profile pages #657
 * September 18 - Obfuscate developer profiles URLs #656